Refactor to create nemo.core.optim
