+++ nemo/core/optim/novograd.py
@@ -12,61 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from argparse import ArgumentParser
-from functools import partial
-
 import torch
-import torch.optim as optim
-from torch.optim import adadelta, adagrad, adamax, rmsprop, rprop
 from torch.optim.optimizer import Optimizer
 
-__all__ = ['Novograd', 'get_optimizer', 'register_optimizer', 'parse_optimizer_args', 'add_optimizer_args']
-
-
-AVAILABLE_OPTIMIZERS = {
-    'sgd': optim.SGD,
-    'adam': optim.Adam,
-    'adamw': optim.AdamW,
-    'adadelta': adadelta.Adadelta,
-    'adamax': adamax.Adamax,
-    'adagrad': adagrad.Adagrad,
-    'rmsprop': rmsprop.RMSprop,
-    'rprop': rprop.Rprop,
-}
-
-
-def _boolify(s):
-    if s == 'True' or s == 'true':
-        return True
-    if s == 'False' or s == 'false':
-        return False
-    raise ValueError('Not Boolean Value!')
-
-
-def _autocast(value):
-    # if value is itself None, dont parse
-    if value is None:
-        return None
-
-    # If value is comma seperated list of items, recursively parse all items in list.
-    if "," in value:
-        values = value.split(',')
-        values = [_autocast(value) for value in values]
-        return values
-
-    # If value is string `none` or `None`, parse as None
-    if value == 'none' or 'None':
-        return None
-
-    # Try type cast and return
-    for cast_type in (int, float, _boolify):
-        try:
-            return cast_type(value)
-        except Exception:
-            pass
-
-    # All types failed, return str without casting
-    return value  # str type
+__all__ = ['Novograd']
 
 
 def _check_valid_opt_params(lr, eps, betas):
@@ -76,86 +25,6 @@ def _check_valid_opt_params(lr, eps, betas):
         raise ValueError(f"Invalid epsilon value: {eps}")
     if not (0.0 <= betas[0] < 1.0 and 0.0 <= betas[1] < 1.0):
         raise ValueError(f"Betas have to be between 0 and 1: {betas}")
-
-
-def parse_optimizer_args(optimizer_kwargs):
-    kwargs = {}
-
-    if optimizer_kwargs is None:
-        return kwargs
-
-    # If it is a pre-defined dictionary, just return its values
-    if hasattr(optimizer_kwargs, 'keys'):
-        return optimizer_kwargs
-
-    # If it is key=value string list, parse all items
-    for key_value in optimizer_kwargs:
-        key, str_value = key_value.split('=')
-
-        value = _autocast(str_value)
-        kwargs[key] = value
-
-    return kwargs
-
-
-def add_optimizer_args(parent_parser: ArgumentParser, optimizer='adam', default_opt_args=None) -> ArgumentParser:
-    """Extends existing argparse with support for optimizers.
-
-    Args:
-        parent_parser (ArgumentParser): Custom CLI parser that will be extended.
-        optimizer (str): Default optimizer required.
-        default_opt_args (list(str)): List of overriding arguments for the instantiated optimizer.
-
-    Returns:
-        ArgumentParser: Parser extended by Optimizers arguments.
-    """
-    if default_opt_args is None:
-        default_opt_args = []
-
-    parser = ArgumentParser(parents=[parent_parser], add_help=True, conflict_handler='resolve')
-
-    parser.add_argument('--optimizer', type=str, default=optimizer, help='Name of the optimizer. Defaults to Adam.')
-    parser.add_argument('--lr', type=float, required=True, help='Learning rate of the optimizer.')
-    parser.add_argument(
-        '--opt_args',
-        default=default_opt_args,
-        nargs='+',
-        type=str,
-        help='Overriding arguments for the optimizer. \n'
-        'Must follow the pattern : \n'
-        'name=value seperated by spaces.',
-    )
-
-    return parser
-
-
-def register_optimizer(name, optimizer: Optimizer):
-    if name in AVAILABLE_OPTIMIZERS:
-        raise ValueError(f"Cannot override pre-existing optimizers. Conflicting optimizer name = {name}")
-
-    AVAILABLE_OPTIMIZERS[name] = optimizer
-
-
-def get_optimizer(name, **kwargs):
-    if name not in AVAILABLE_OPTIMIZERS:
-        raise ValueError(
-            f"Cannot resolve optimizer '{name}'. Available optimizers are : " f"{AVAILABLE_OPTIMIZERS.keys()}"
-        )
-
-    optimizer = AVAILABLE_OPTIMIZERS[name]
-    optimizer = partial(optimizer, **kwargs)
-    return optimizer
-
-
-def master_params(optimizer):
-    """
-    Generator expression that iterates over the params owned by ``optimizer``.
-    Args:
-        optimizer: An optimizer previously returned from ``amp.initialize``.
-    """
-    for group in optimizer.param_groups:
-        for p in group['params']:
-            yield p
 
 
 class Novograd(Optimizer):
@@ -274,7 +143,3 @@ class Novograd(Optimizer):
                     p.data.add_(-group["lr"], exp_avg)
 
         return loss
-
-
-# Register Novograd
-register_optimizer('novograd', Novograd)
